requirements.txt
amqp==2.5.2
attrs==19.1.0
billiard==3.6.1.0
boto3==1.10.31
botocore==1.13.31
celery==4.4.0
certifi==2019.6.16
chardet==3.0.4
coolname==1.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.6
django-appconf==1.0.3
django-choices==1.7.1
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-filter==2.1.0
django-ipware==2.1.0
django-storages==1.8
djangorestframework==3.9.4
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.16.0
googlemaps==4.4.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.4.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
Pillow==6.1.0
pinax-stripe==4.4.0
psycopg2-binary==2.8.3
pygeocoder==1.2.5
PyJWT==1.7.1
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3transfer==0.2.1
sentry-sdk==0.12.3
six==1.15.0
sqlparse==0.3.0
stripe==2.32.1
twilio==6.29.4
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.4
zipp==2.0.1
